Updated all environments to support the new way of dealing with time:
Time is unsigned long, if this is not good enough then the environment is
free to define ENV_TIME_T and supply their own time (take a look at env/types.h)
for how this is done in the standard case.

Also small fixes to the env/env.h file, no longer thinks that SRP is a special
case for an environment but rather treat them as a normal environment.